core: boot: make boot parameters unsigned long
Changes all boot parameters passed in registers to be of the type
unsigned long to match the register size and allow greater width on
64-bit systems.

Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU v7 & v8)

+++ core/arch/arm/include/kernel/generic_boot.h
@@ -30,19 +30,19 @@
 #include <stdint.h>
 
 #if defined(CFG_WITH_ARM_TRUSTED_FW)
-uint32_t cpu_on_handler(uint32_t a0, uint32_t a1);
-uint32_t *generic_boot_init_primary(uint32_t pageable_part, uint32_t unused,
-				    uint32_t fdt);
-uint32_t generic_boot_cpu_on_handler(uint32_t a0, uint32_t a1);
+unsigned long cpu_on_handler(unsigned long a0, unsigned long a1);
+uint32_t *generic_boot_init_primary(unsigned long pageable_part,
+				    unsigned long unused, unsigned long fdt);
+unsigned long generic_boot_cpu_on_handler(unsigned long a0, unsigned long a1);
 #else
-void generic_boot_init_primary(uint32_t pageable_part, uint32_t nsec_entry,
-			       uint32_t fdt);
-void generic_boot_init_secondary(uint32_t nsec_entry);
+void generic_boot_init_primary(unsigned long pageable_part,
+			       unsigned long nsec_entry, unsigned long fdt);
+void generic_boot_init_secondary(unsigned long nsec_entry);
 #endif
 
 void main_init_gic(void);
 
-void init_sec_mon(uint32_t nsec_entry);
+void init_sec_mon(unsigned long nsec_entry);
 
 const struct thread_handlers *generic_boot_get_handlers(void);
